+++ modules/uefi_drivers/uefi_vmm/walk.h
@@ -15,7 +15,7 @@ enum walk_error {
 #define WALK_2MB 0x200000
 #define WALK_1GB 0x40000000
 
-uint64_t walk_long(uint64_t cr3, uint64_t linear, uint64_t *physical, uint64_t
+int walk_long(uint64_t cr3, uint64_t linear, uint64_t *physical, uint64_t
     *size);
 
 uint8_t walk_is_in_page(uint64_t l, uint64_t p, uint64_t size);
